requirement.txt
addict
argon2-cffi
argon2-cffi-bindings
asttokens
attrs
backcall
black
bleach
cachetools
certifi
cffi
click
cycler
debugpy
decorator
defusedxml
descartes
efficientnet-pytorch
entrypoints
executing
fire
fonttools
importlib-resources
ipdb
ipykernel
ipython
ipython-genutils
ipywidgets
jedi
Jinja2
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kiwisolver
MarkupSafe
matplotlib
matplotlib-inline
mistune
mmcv
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpy
nuscenes-devkit
opencv-python
packaging
pandocfilters
parso
pathspec
pexpect
pickleshare
Pillow
platformdirs
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pure-eval
pycocotools
pycparser
Pygments
pyparsing
pyquaternion
pyrsistent
python-dateutil
PyYAML
pyzmq
qtconsole
QtPy
scikit-learn
scipy
Send2Trash
Shapely==1.8.1.post1
six==1.16.0
stack-data
tensorboardX==2.4.1
termcolor
terminado
testpath
threadpoolctl
toml
tomli
torch-scatter

